bnxt: set the VMDQ pool size correctly
Calculate the VMDQ pool size correctly:
For PFs the VMDQ count needs to be the minimum of all used and available
resources like vnic, l2 CTX, RSS CTX, or 64 (hard coded).
For VFs restrict the number of VMDQs to 1.
If VFs are enabled, only allocate a single vnic to each function.
This allows targeting a VF with a single L2 context.

In the future, we will enable VMDq on VFs.

We are essentially mapping VMDq pools to VNICs now.  If there are
no VFs, the PF can use all VNICs as VMDq pools.  If there are VFs,
each function (including the PF) will get a single VNIC.

Signed-off-by: Stephen Hurd <stephen.hurd@broadcom.com>
Signed-off-by: Ajit Khaparde <ajit.khaparde@broadcom.com>