bnxt: add support for PF/VF communications
Set up interrupts and default completion ring during device init to
allow VF communications.

Request most VF HWRM commands be forwarded to the PF driver.  In the
future, this should be configurable.

Fix reporting of active VFs.

Add a log message when a completion is ignored on the default completion
ring since this will indicate support that needs to be added.

Do not initialize the default completion ring in bnxt_alloc_hwrm_rings(),
but move it out to bnxt_dev_init().

Create a structure bnxt_child_vf_info for PF to cache VF specific info.

Signed-off-by: Stephen Hurd <stephen.hurd@broadcom.com>
Signed-off-by: Ajit Khaparde <ajit.khaparde@broadcom.com>