bnxt: fix issues with completion ring processing
The current completion ring processing is buggy.
Completion doorbell index needs to be the last consumed, not next free.
In few cases this causes a handful of completions to be processed.
The remaining completions could be stuck waiting for processing.
This patch tries to avoid that. To do that, we track the index directly
rather than indirectly via raw_cons.
Now, we explicitly register the async completion ring with the firmware.

Signed-off-by: Stephen Hurd <stephen.hurd@broadcom.com>
Signed-off-by: Ajit Khaparde <ajit.khaparde@broadcom.com>