net/bnxt: reuse Rx mbufs
Support Rx mbuf reuse.
Allow the PMD to allocate the Rx ring with mbufs and then reuse
the allocated buffers after every packet use.

The application should ensure that the mbuf is consumed before
the NIC hardware completes the DMA of a new packet at the same
mbuf location.

The feature can help reduce the amount of time needed to allocate
and free mbufs for each packet.

The user can enable this functionality using the devarg
rx-mbuf-reuse.

Signed-off-by: Ajit Khaparde <ajit.khaparde@broadcom.com>